backend/requirements.txt
# ============================================================================
# SmartBudget Backend Requirements
# ============================================================================
# Python 3.13.8+
# Updated: December 3, 2025
# Status: All dependencies audited and verified for application architecture
# ============================================================================

# ============================================================================
# CORE WEB FRAMEWORK & SERVER
# ============================================================================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.6

# ============================================================================
# CONFIGURATION & ENVIRONMENT
# ============================================================================
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.0.3

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy==2.0.23

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=46.0.0

# ============================================================================
# API INTEGRATION - BANKING & FINANCIAL DATA
# ============================================================================
plaid-python==12.0.0
aiohttp>=3.13.0

# ============================================================================
# DATA VALIDATION & PARSING
# ============================================================================
email-validator==2.1.0

# ============================================================================
# DATA ANALYSIS & NUMERICAL COMPUTING
# ============================================================================
pandas==2.3.3
numpy==2.3.5
openpyxl==3.1.5

# ============================================================================
# SEC FILING & FINANCIAL DOCUMENT ANALYSIS
# ============================================================================
sec-edgar-downloader>=5.0.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
requests>=2.31.0

# ============================================================================
# HTTP CLIENT & TESTING
# ============================================================================
httpx>=0.27.0,<0.28.0
